+++ controller/backend/AdminUserController.php
@@ -16,7 +16,7 @@ Class AdminUserController
 		}
 		else
 		{
-			header('Location: index.php?action=admin');
+			header('Location: ../admin/');
 		}	
 	}
 
@@ -24,29 +24,47 @@ Class AdminUserController
 	{
 		$userManager = new Cornelissen\Shapeplace\Model\UserManager();
 
-		$give = $userManager->givePower($idUser);
-
-		$_SESSION['success'] = 'Les droits sont mis en place.';
-		header('Location: index.php?action=adminUser');
+		if($_SESSION['role'] == 'superAdmin')
+		{
+			$give = $userManager->givePower($idUser);
+			$_SESSION['success'] = 'Les droits sont mis en place.';
+			header('Location: ../adminUser/');
+		}
+		else
+		{
+			header('Location: ../admin/');
+		}	
 	}
 
 	public function remainPower($idUser)
 	{
 		$userManager = new Cornelissen\Shapeplace\Model\UserManager();
 
-		$give = $userManager->remainPower($idUser);
-
-		$_SESSION['success'] = 'Les droits sont enlevé.';
-		header('Location: index.php?action=adminUser');
+		if($_SESSION['role'] == 'superAdmin')
+		{
+			$give = $userManager->remainPower($idUser);
+			$_SESSION['success'] = 'Les droits sont enlevé.';
+			header('Location: ../adminUser/');
+		}
+		else
+		{
+			header('Location: ../admin/');
+		}	
 	}
 
 	public function deleteUser($idUser)
 	{
 		$userManager = new Cornelissen\Shapeplace\Model\UserManager();
 
-		$give = $userManager->deleteUser($idUser);
-
-		$_SESSION['success'] = 'L\'utilisateur est supprimé.';
-		header('Location: index.php?action=adminUser');
+		if($_SESSION['role'] == 'superAdmin')
+		{
+			$give = $userManager->deleteUser($idUser);
+			$_SESSION['success'] = 'L\'utilisateur est supprimé.';
+			header('Location: ../adminUser/');
+		}
+		else
+		{
+			header('Location: ../admin/');
+		}
 	}
 }
